pyproject.toml
[tool.poetry]
name = "truss"
version = "0.0.20"
description = ""
authors = ["Alex Gillmor <alex@baseten.co>"]
include = ["*.txt", "*.Dockerfile", "*.md"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
numpy = "^1.18"
msgpack = "1.0.2"
msgpack-numpy = "0.4.7.1"
packaging = "20.9"
python-json-logger = "2.0.2"
PyYAML = "^6.0"
pre-commit = "^2.18.1"
Jinja2 = "^3.1.2"
python-on-whales = "^0.46.0"
tenacity = "^8.0.1"

[tool.poetry.dev-dependencies]
torch = "^1.9.0"
ipython = "^7.16"
pytest = "6.2.4"
tensorflow = "^2.4.0"
transformers = "<=4.18.0"
scikit-learn = "1.0.2"
pandas = "1.1.5"
tensorflow-hub = "^0.12.0"
kfserving = "^0.3"
isort = "^5.10.1"
flake8 = "^4.0.1"
ipdb = "^0.13.9"
coverage = "^6.4.1"
pytest-cov = "^3.0.0"

[tool.poetry.scripts]
truss = 'truss.cli:cli_group'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
